change component to login
